:evergreen_tree: Convert tree weights to floats for easier v0 compat

+++ tree/node.go
@@ -15,7 +15,7 @@ type Node struct {
 	Brain     SnakeBrain
 
 	Depth  int
-	Weight int32
+	Weight float64
 }
 
 func (n *Node) String() string {
